requirements.txt
# Core Dependencies
dspy-ai>=2.4.0
openai>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
img2table>=1.2.0

# Excel Processing
openpyxl>=3.1.0
pandas>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.0
regex>=2023.0.0

# CLI
click>=8.1.0
colorama>=0.4.6
